mqtt_gateway/requirements.txt
paho-mqtt==2.1.0
numpy==1.25.2
python-dotenv==0.19.0 
pydantic==2.7.1
